[tokenizer] Split TranslatorTest into smaller test classes.
