--- extensions/tokenizers/src/test/java/ai/djl/huggingface/tokenizers/TranslatorTest.java
@@ -1,609 +1,0 @@
-/*
- * Copyright 2022 Amazon.com, Inc. or its affiliates. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"). You may not use this file except in compliance
- * with the License. A copy of the License is located at
- *
- * http://aws.amazon.com/apache2.0/
- *
- * or in the "license" file accompanying this file. This file is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES
- * OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions
- * and limitations under the License.
- */
-package ai.djl.huggingface.tokenizers;
-
-import ai.djl.Model;
-import ai.djl.ModelException;
-import ai.djl.huggingface.translator.FillMaskTranslatorFactory;
-import ai.djl.huggingface.translator.QuestionAnsweringTranslatorFactory;
-import ai.djl.huggingface.translator.TextClassificationTranslatorFactory;
-import ai.djl.huggingface.translator.TextEmbeddingTranslatorFactory;
-import ai.djl.huggingface.translator.TokenClassificationTranslatorFactory;
-import ai.djl.inference.Predictor;
-import ai.djl.modality.Classifications;
-import ai.djl.modality.Input;
-import ai.djl.modality.Output;
-import ai.djl.modality.nlp.qa.QAInput;
-import ai.djl.modality.nlp.translator.NamedEntity;
-import ai.djl.ndarray.NDArray;
-import ai.djl.ndarray.NDList;
-import ai.djl.ndarray.NDManager;
-import ai.djl.ndarray.types.Shape;
-import ai.djl.nn.Block;
-import ai.djl.nn.LambdaBlock;
-import ai.djl.repository.zoo.Criteria;
-import ai.djl.repository.zoo.ZooModel;
-import ai.djl.testing.Assertions;
-import ai.djl.translate.TranslateException;
-import ai.djl.util.JsonUtils;
-import ai.djl.util.Utils;
-
-import org.testng.Assert;
-import org.testng.annotations.AfterClass;
-import org.testng.annotations.Test;
-
-import java.io.IOException;
-import java.io.Writer;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.util.HashMap;
-import java.util.Map;
-
-public class TranslatorTest {
-
-    @AfterClass
-    public void tierDown() {
-        Utils.deleteQuietly(Paths.get("build/model"));
-    }
-
-    @Test
-    public void testQATranslator() throws ModelException, IOException, TranslateException {
-        String question = "When did BBC Japan start broadcasting?";
-        String paragraph =
-                "BBC Japan was a general entertainment Channel. "
-                        + "Which operated between December 2004 and April 2006. "
-                        + "It ceased operations after its Japanese distributor folded.";
-
-        Block block =
-                new LambdaBlock(
-                        a -> {
-                            NDManager manager = a.getManager();
-                            long[][] start = new long[1][36];
-                            long[][] end = new long[1][36];
-                            start[0][0] = 2;
-                            start[0][21] = 1;
-                            end[0][0] = 2;
-                            end[0][20] = 1;
-                            NDArray arr1 = manager.create(start);
-                            NDArray arr2 = manager.create(end);
-                            return new NDList(arr1, arr2);
-                        },
-                        "model");
-        Path modelDir = Paths.get("build/model");
-        Files.createDirectories(modelDir);
-
-        Criteria<QAInput, String> criteria =
-                Criteria.builder()
-                        .setTypes(QAInput.class, String.class)
-                        .optModelPath(modelDir)
-                        .optBlock(block)
-                        .optEngine("PyTorch")
-                        .optArgument("tokenizer", "bert-base-cased")
-                        .optOption("hasParameter", "false")
-                        .optTranslatorFactory(new QuestionAnsweringTranslatorFactory())
-                        .build();
-
-        try (ZooModel<QAInput, String> model = criteria.loadModel();
-                Predictor<QAInput, String> predictor = model.newPredictor()) {
-            QAInput input = new QAInput(question, paragraph);
-            String res = predictor.predict(input);
-            Assert.assertEquals(res, "December 2004");
-        }
-
-        Criteria<Input, Output> criteria2 =
-                Criteria.builder()
-                        .setTypes(Input.class, Output.class)
-                        .optModelPath(modelDir)
-                        .optBlock(block)
-                        .optEngine("PyTorch")
-                        .optArgument("tokenizer", "bert-base-cased")
-                        .optOption("hasParameter", "false")
-                        .optTranslatorFactory(new QuestionAnsweringTranslatorFactory())
-                        .build();
-
-        try (ZooModel<Input, Output> model = criteria2.loadModel();
-                Predictor<Input, Output> predictor = model.newPredictor()) {
-            Input input = new Input();
-            input.add("question", question);
-            input.add("paragraph", paragraph);
-            Output res = predictor.predict(input);
-            Assert.assertEquals(res.getAsString(0), "December 2004");
-        }
-
-        try (Model model = Model.newInstance("test")) {
-            model.setBlock(block);
-            Map<String, String> options = new HashMap<>();
-            options.put("hasParameter", "false");
-            model.load(modelDir, "test", options);
-
-            QuestionAnsweringTranslatorFactory factory = new QuestionAnsweringTranslatorFactory();
-            Map<String, String> arguments = new HashMap<>();
-
-            Assert.assertThrows(
-                    TranslateException.class,
-                    () -> factory.newInstance(String.class, Integer.class, model, arguments));
-
-            arguments.put("tokenizer", "bert-base-cased");
-
-            Assert.assertThrows(
-                    IllegalArgumentException.class,
-                    () -> factory.newInstance(String.class, Integer.class, model, arguments));
-        }
-    }
-
-    @Test
-    public void testFillMaskTranslator() throws ModelException, IOException, TranslateException {
-        String text = "Hello I'm a [MASK] model.";
-
-        Block block =
-                new LambdaBlock(
-                        a -> {
-                            NDManager manager = a.getManager();
-                            float[][] logits = new float[10][4828];
-                            logits[6][4827] = 5;
-                            logits[6][2535] = 4;
-                            logits[6][2047] = 3;
-                            logits[6][3565] = 2;
-                            logits[6][2986] = 1;
-                            NDArray arr = manager.create(logits);
-                            arr = arr.expandDims(0);
-                            return new NDList(arr);
-                        },
-                        "model");
-        Path modelDir = Paths.get("build/model");
-        Files.createDirectories(modelDir);
-
-        Criteria<String, Classifications> criteria =
-                Criteria.builder()
-                        .setTypes(String.class, Classifications.class)
-                        .optModelPath(modelDir)
-                        .optBlock(block)
-                        .optEngine("PyTorch")
-                        .optArgument("tokenizer", "bert-base-uncased")
-                        .optOption("hasParameter", "false")
-                        .optTranslatorFactory(new FillMaskTranslatorFactory())
-                        .build();
-
-        try (ZooModel<String, Classifications> model = criteria.loadModel();
-                Predictor<String, Classifications> predictor = model.newPredictor()) {
-            Classifications res = predictor.predict(text);
-            Assert.assertEquals(res.best().getClassName(), "fashion");
-            Assert.assertThrows(
-                    TranslateException.class,
-                    () -> predictor.predict("Hello I'm a invalid model."));
-            Assert.assertThrows(
-                    TranslateException.class,
-                    () -> predictor.predict("I'm a [MASK] [MASK] model."));
-        }
-
-        Criteria<Input, Output> criteria2 =
-                Criteria.builder()
-                        .setTypes(Input.class, Output.class)
-                        .optModelPath(modelDir)
-                        .optBlock(block)
-                        .optEngine("PyTorch")
-                        .optArgument("tokenizer", "bert-base-uncased")
-                        .optOption("hasParameter", "false")
-                        .optTranslatorFactory(new FillMaskTranslatorFactory())
-                        .build();
-
-        try (ZooModel<Input, Output> model = criteria2.loadModel();
-                Predictor<Input, Output> predictor = model.newPredictor()) {
-            Input input = new Input();
-            input.add(text);
-            Output out = predictor.predict(input);
-            Classifications res = (Classifications) out.getData();
-            Assert.assertEquals(res.best().getClassName(), "fashion");
-        }
-
-        try (Model model = Model.newInstance("test")) {
-            model.setBlock(block);
-            Map<String, String> options = new HashMap<>();
-            options.put("hasParameter", "false");
-            model.load(modelDir, "test", options);
-
-            FillMaskTranslatorFactory factory = new FillMaskTranslatorFactory();
-            Map<String, String> arguments = new HashMap<>();
-
-            Assert.assertThrows(
-                    TranslateException.class,
-                    () -> factory.newInstance(String.class, Integer.class, model, arguments));
-
-            arguments.put("tokenizer", "bert-base-uncased");
-
-            Assert.assertThrows(
-                    IllegalArgumentException.class,
-                    () -> factory.newInstance(String.class, Integer.class, model, arguments));
-        }
-    }
-
-    @Test
-    public void testTokenClassificationTranslator()
-            throws ModelException, IOException, TranslateException {
-        String text = "My name is Wolfgang and I live in Berlin.";
-
-        Block block =
-                new LambdaBlock(
-                        a -> {
-                            NDManager manager = a.getManager();
-                            float[][] logits = new float[12][9];
-                            logits[4][3] = 1;
-                            logits[9][7] = 1;
-                            NDArray arr = manager.create(logits);
-                            arr = arr.expandDims(0);
-                            return new NDList(arr);
-                        },
-                        "model");
-        Path modelDir = Paths.get("build/model");
-        Files.createDirectories(modelDir);
-        Path path = modelDir.resolve("config.json");
-        Map<String, Map<String, String>> map = new HashMap<>();
-        Map<String, String> id2label = new HashMap<>();
-        id2label.put("0", "O");
-        id2label.put("3", "B-PER");
-        id2label.put("7", "B-LOC");
-        map.put("id2label", id2label);
-        try (Writer writer = Files.newBufferedWriter(path)) {
-            writer.write(JsonUtils.GSON.toJson(map));
-        }
-
-        Criteria<String, NamedEntity[]> criteria =
-                Criteria.builder()
-                        .setTypes(String.class, NamedEntity[].class)
-                        .optModelPath(modelDir)
-                        .optBlock(block)
-                        .optEngine("PyTorch")
-                        .optArgument("tokenizer", "bert-base-uncased")
-                        .optOption("hasParameter", "false")
-                        .optTranslatorFactory(new TokenClassificationTranslatorFactory())
-                        .build();
-
-        try (ZooModel<String, NamedEntity[]> model = criteria.loadModel();
-                Predictor<String, NamedEntity[]> predictor = model.newPredictor()) {
-            NamedEntity[] res = predictor.predict(text);
-            Assert.assertEquals(res[0].getEntity(), "B-PER");
-            Assertions.assertAlmostEquals(res[0].getScore(), 0.2536117);
-            Assert.assertEquals(res[0].getIndex(), 4);
-            Assert.assertEquals(res[0].getWord(), "wolfgang");
-            Assert.assertEquals(res[0].getStart(), 11);
-            Assert.assertEquals(res[0].getEnd(), 19);
-        }
-
-        Criteria<Input, Output> criteria2 =
-                Criteria.builder()
-                        .setTypes(Input.class, Output.class)
-                        .optModelPath(modelDir)
-                        .optBlock(block)
-                        .optEngine("PyTorch")
-                        .optArgument("tokenizer", "bert-base-uncased")
-                        .optOption("hasParameter", "false")
-                        .optTranslatorFactory(new TokenClassificationTranslatorFactory())
-                        .build();
-
-        try (ZooModel<Input, Output> model = criteria2.loadModel();
-                Predictor<Input, Output> predictor = model.newPredictor()) {
-            Input input = new Input();
-            input.add(text);
-            Output out = predictor.predict(input);
-            NamedEntity[] res = (NamedEntity[]) out.getData().getAsObject();
-            Assert.assertEquals(res[0].getEntity(), "B-PER");
-        }
-
-        try (Model model = Model.newInstance("test")) {
-            model.setBlock(block);
-            Map<String, String> options = new HashMap<>();
-            options.put("hasParameter", "false");
-            model.load(modelDir, "test", options);
-
-            TokenClassificationTranslatorFactory factory =
-                    new TokenClassificationTranslatorFactory();
-            Map<String, String> arguments = new HashMap<>();
-
-            Assert.assertThrows(
-                    TranslateException.class,
-                    () -> factory.newInstance(String.class, Integer.class, model, arguments));
-
-            arguments.put("tokenizer", "bert-base-uncased");
-
-            Assert.assertThrows(
-                    IllegalArgumentException.class,
-                    () -> factory.newInstance(String.class, Integer.class, model, arguments));
-        }
-    }
-
-    @Test
-    public void testTextEmbeddingTranslator()
-            throws ModelException, IOException, TranslateException {
-        String text = "This is an example sentence";
-
-        Block block =
-                new LambdaBlock(
-                        a -> {
-                            NDManager manager = a.getManager();
-                            NDArray arr = manager.ones(new Shape(1, 7, 384));
-                            arr.setName("last_hidden_state");
-                            return new NDList(arr);
-                        },
-                        "model");
-        Path modelDir = Paths.get("build/model");
-        Files.createDirectories(modelDir);
-
-        Criteria<String, float[]> criteria =
-                Criteria.builder()
-                        .setTypes(String.class, float[].class)
-                        .optModelPath(modelDir)
-                        .optBlock(block)
-                        .optEngine("PyTorch")
-                        .optArgument("tokenizer", "bert-base-uncased")
-                        .optOption("hasParameter", "false")
-                        .optTranslatorFactory(new TextEmbeddingTranslatorFactory())
-                        .build();
-
-        try (ZooModel<String, float[]> model = criteria.loadModel();
-                Predictor<String, float[]> predictor = model.newPredictor()) {
-            float[] res = predictor.predict(text);
-            Assert.assertEquals(res.length, 384);
-            Assertions.assertAlmostEquals(res[0], 0.05103);
-        }
-
-        // pooling_mode_max_tokens
-        criteria =
-                Criteria.builder()
-                        .setTypes(String.class, float[].class)
-                        .optModelPath(modelDir)
-                        .optBlock(block)
-                        .optEngine("PyTorch")
-                        .optArgument("tokenizer", "bert-base-uncased")
-                        .optArgument("pooling", "max")
-                        .optOption("hasParameter", "false")
-                        .optTranslatorFactory(new TextEmbeddingTranslatorFactory())
-                        .build();
-
-        try (ZooModel<String, float[]> model = criteria.loadModel();
-                Predictor<String, float[]> predictor = model.newPredictor()) {
-            float[] res = predictor.predict(text);
-            Assert.assertEquals(res.length, 384);
-            Assertions.assertAlmostEquals(res[0], 1.0);
-        }
-
-        // pooling_mean_sqrt_len_tokens
-        criteria =
-                Criteria.builder()
-                        .setTypes(String.class, float[].class)
-                        .optModelPath(modelDir)
-                        .optBlock(block)
-                        .optEngine("PyTorch")
-                        .optArgument("tokenizer", "bert-base-uncased")
-                        .optArgument("pooling", "mean_sqrt_len")
-                        .optOption("hasParameter", "false")
-                        .optTranslatorFactory(new TextEmbeddingTranslatorFactory())
-                        .build();
-
-        try (ZooModel<String, float[]> model = criteria.loadModel();
-                Predictor<String, float[]> predictor = model.newPredictor()) {
-            float[] res = predictor.predict(text);
-            Assert.assertEquals(res.length, 384);
-            Assertions.assertAlmostEquals(res[0], 0.05103104);
-        }
-
-        // pooling_weightedmean_tokens
-        criteria =
-                Criteria.builder()
-                        .setTypes(String.class, float[].class)
-                        .optModelPath(modelDir)
-                        .optBlock(block)
-                        .optEngine("PyTorch")
-                        .optArgument("tokenizer", "bert-base-uncased")
-                        .optArgument("pooling", "weightedmean")
-                        .optOption("hasParameter", "false")
-                        .optTranslatorFactory(new TextEmbeddingTranslatorFactory())
-                        .build();
-
-        try (ZooModel<String, float[]> model = criteria.loadModel();
-                Predictor<String, float[]> predictor = model.newPredictor()) {
-            float[] res = predictor.predict(text);
-            Assert.assertEquals(res.length, 384);
-            Assertions.assertAlmostEquals(res[0], 0.05103104);
-        }
-
-        Criteria<Input, Output> criteria2 =
-                Criteria.builder()
-                        .setTypes(Input.class, Output.class)
-                        .optModelPath(modelDir)
-                        .optBlock(block)
-                        .optEngine("PyTorch")
-                        .optArgument("tokenizer", "bert-base-uncased")
-                        .optArgument("pooling", "cls")
-                        .optOption("hasParameter", "false")
-                        .optTranslatorFactory(new TextEmbeddingTranslatorFactory())
-                        .build();
-
-        try (ZooModel<Input, Output> model = criteria2.loadModel();
-                Predictor<Input, Output> predictor = model.newPredictor()) {
-            Input input = new Input();
-            input.add(text);
-            Output out = predictor.predict(input);
-            float[] res = JsonUtils.GSON.fromJson(out.getAsString(0), float[].class);
-            Assert.assertEquals(res.length, 384);
-            Assertions.assertAlmostEquals(res[0], 0.05103);
-        }
-
-        try (Model model = Model.newInstance("test")) {
-            model.setBlock(block);
-            Map<String, String> options = new HashMap<>();
-            options.put("hasParameter", "false");
-            model.load(modelDir, "test", options);
-
-            TextEmbeddingTranslatorFactory factory = new TextEmbeddingTranslatorFactory();
-            Map<String, String> arguments = new HashMap<>();
-
-            Assert.assertThrows(
-                    TranslateException.class,
-                    () -> factory.newInstance(String.class, Integer.class, model, arguments));
-
-            arguments.put("tokenizer", "bert-base-uncased");
-
-            Assert.assertThrows(
-                    IllegalArgumentException.class,
-                    () -> factory.newInstance(String.class, Integer.class, model, arguments));
-        }
-    }
-
-    @Test
-    public void testTextEmbeddingBatchTranslator()
-            throws ModelException, IOException, TranslateException {
-        String[] text = {"This is an example sentence", "This is the second sentence"};
-
-        Block block =
-                new LambdaBlock(
-                        a -> {
-                            NDManager manager = a.getManager();
-                            NDArray arr = manager.ones(new Shape(2, 7, 384));
-                            arr.setName("last_hidden_state");
-                            return new NDList(arr);
-                        },
-                        "model");
-        Path modelDir = Paths.get("build/model");
-        Files.createDirectories(modelDir);
-
-        Criteria<String[], float[][]> criteria =
-                Criteria.builder()
-                        .setTypes(String[].class, float[][].class)
-                        .optModelPath(modelDir)
-                        .optBlock(block)
-                        .optEngine("PyTorch")
-                        .optArgument("tokenizer", "bert-base-uncased")
-                        .optArgument("padding", "true")
-                        .optOption("hasParameter", "false")
-                        .optTranslatorFactory(new TextEmbeddingTranslatorFactory())
-                        .build();
-
-        try (ZooModel<String[], float[][]> model = criteria.loadModel();
-                Predictor<String[], float[][]> predictor = model.newPredictor()) {
-            float[][] res = predictor.predict(text);
-            Assert.assertEquals(res[0].length, 384);
-            Assertions.assertAlmostEquals(res[0][0], 0.05103);
-        }
-
-        Criteria<Input, Output> criteria2 =
-                Criteria.builder()
-                        .setTypes(Input.class, Output.class)
-                        .optModelPath(modelDir)
-                        .optBlock(block)
-                        .optEngine("PyTorch")
-                        .optArgument("tokenizer", "bert-base-uncased")
-                        .optArgument("padding", "true")
-                        .optOption("hasParameter", "false")
-                        .optTranslatorFactory(new TextEmbeddingTranslatorFactory())
-                        .build();
-
-        try (ZooModel<Input, Output> model = criteria2.loadModel();
-                Predictor<Input, Output> predictor = model.newPredictor()) {
-            Input input = new Input();
-            input.add(JsonUtils.GSON.toJson(text));
-            input.addProperty("Content-Type", "application/json");
-            Output out = predictor.predict(input);
-            float[][] res = JsonUtils.GSON.fromJson(out.getAsString(0), float[][].class);
-            Assert.assertEquals(res[0].length, 384);
-            Assertions.assertAlmostEquals(res[0][0], 0.05103);
-        }
-    }
-
-    @Test
-    public void testTextClassificationTranslator()
-            throws ModelException, IOException, TranslateException {
-        String text = "DJL is the best.";
-
-        Block block =
-                new LambdaBlock(
-                        a -> {
-                            NDManager manager = a.getManager();
-                            float[] logits = new float[] {0.02f, 0.2f, 0.97f};
-                            NDArray arr = manager.create(logits, new Shape(1, 3));
-                            arr.setName("logits");
-                            return new NDList(arr);
-                        },
-                        "model");
-        Path modelDir = Paths.get("build/model");
-        Files.createDirectories(modelDir);
-
-        Path path = modelDir.resolve("config.json");
-        Map<String, Map<String, String>> map = new HashMap<>();
-        Map<String, String> id2label = new HashMap<>();
-        id2label.put("0", "LABEL_0");
-        id2label.put("1", "LABEL_1");
-        id2label.put("2", "LABEL_2");
-        map.put("id2label", id2label);
-        try (Writer writer = Files.newBufferedWriter(path)) {
-            writer.write(JsonUtils.GSON.toJson(map));
-        }
-
-        Criteria<String, Classifications> criteria =
-                Criteria.builder()
-                        .setTypes(String.class, Classifications.class)
-                        .optModelPath(modelDir)
-                        .optBlock(block)
-                        .optEngine("PyTorch")
-                        .optArgument("tokenizer", "bert-base-uncased")
-                        .optOption("hasParameter", "false")
-                        .optTranslatorFactory(new TextClassificationTranslatorFactory())
-                        .build();
-
-        try (ZooModel<String, Classifications> model = criteria.loadModel();
-                Predictor<String, Classifications> predictor = model.newPredictor()) {
-            Classifications res = predictor.predict(text);
-            Assert.assertEquals(res.best().getClassName(), "LABEL_2");
-        }
-
-        Criteria<Input, Output> criteria2 =
-                Criteria.builder()
-                        .setTypes(Input.class, Output.class)
-                        .optModelPath(modelDir)
-                        .optBlock(block)
-                        .optEngine("PyTorch")
-                        .optArgument("tokenizer", "bert-base-uncased")
-                        .optOption("hasParameter", "false")
-                        .optTranslatorFactory(new TextClassificationTranslatorFactory())
-                        .build();
-
-        try (ZooModel<Input, Output> model = criteria2.loadModel();
-                Predictor<Input, Output> predictor = model.newPredictor()) {
-            Input input = new Input();
-            input.add(text);
-            Output out = predictor.predict(input);
-            Classifications res = (Classifications) out.getData();
-            Assert.assertEquals(res.best().getClassName(), "LABEL_2");
-        }
-
-        try (Model model = Model.newInstance("test")) {
-            model.setBlock(block);
-            Map<String, String> options = new HashMap<>();
-            options.put("hasParameter", "false");
-            model.load(modelDir, "test", options);
-
-            TextClassificationTranslatorFactory factory = new TextClassificationTranslatorFactory();
-            Map<String, String> arguments = new HashMap<>();
-
-            Assert.assertThrows(
-                    TranslateException.class,
-                    () -> factory.newInstance(String.class, Integer.class, model, arguments));
-
-            arguments.put("tokenizer", "bert-base-uncased");
-
-            Assert.assertThrows(
-                    IllegalArgumentException.class,
-                    () -> factory.newInstance(String.class, Integer.class, model, arguments));
-        }
-    }
-}
